+++ src/app/routes/connector-ui/asset-page/asset-edit-dialog/asset-edit-dialog-data.ts
@@ -1,0 +1,5 @@
+import {AssetEditorDialogFormValue} from './form/model/asset-editor-dialog-form-model';
+
+export interface AssetEditDialogData {
+  initialFormValue: AssetEditorDialogFormValue;
+}